ProductTest - prepare to pull request
